Add files via upload
a game where you have to get 15 fruits in the bowl before the timer of 30sec runs out! players use the left and right keys to move the bowl, and if they get less then 0 fruits or the timer runs out the game is over. but if they get 15 fruit in the bowl they get a congratulation screen!